+++ src/device/colorbulb.ts
@@ -325,7 +325,7 @@ export class ColorBulb {
               this.infoLog(`Color Bulb: ${this.accessory.displayName} BLE Address Found: ${this.address}`);
               this.infoLog(`Color Bulb: ${this.accessory.displayName} Config BLE Address: ${this.device.bleMac}`);
             }
-            this.errorLog(`Color Bulb: ${this.accessory.displayName} serviceData: ${JSON.stringify(ad.serviceData)}`);
+            this.debugLog(`Color Bulb: ${this.accessory.displayName} serviceData: ${JSON.stringify(ad.serviceData)}`);
             this.serviceData = ad.serviceData;
             //this.state = ad.serviceData.state;
             //this.delay = ad.serviceData.delay;